category and product service implemented

+++ src/main/java/com/webshop/cart/CartDTO.java
@@ -1,10 +1,13 @@
 package com.webshop.cart;
 
+import com.webshop.cartItem.CartItemDTO;
+import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -13,6 +16,6 @@ import java.util.List;
 @NoArgsConstructor
 public class CartDTO {
     private List<CartItemDTO> items;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 }
